Rename importable module
