+++ sherlock_project/notify.py
@@ -3,7 +3,7 @@
 This module defines the objects for notifying the caller about the
 results of queries.
 """
-from sherlock.result import QueryStatus
+from sherlock_project.result import QueryStatus
 from colorama import Fore, Style
 import webbrowser
