+++ tests/test_ux.py
@@ -1,5 +1,5 @@
 import pytest
-from sherlock import sherlock
+from sherlock_project import sherlock
 from sherlock_interactives import Interactives
 from sherlock_interactives import InteractivesSubprocessError
